requirements.txt
python-telegram-bot==22.7
yt-dlp